+++ specs/Qowaiv.Validation.Specs/DataAnnotations/Stubs/ServiceProviderStub.cs
@@ -1,0 +1,6 @@
+﻿namespace Specs.DataAnnotations.Subs;
+
+internal sealed class ServiceProviderStub : Dictionary<Type, object>, IServiceProvider
+{
+    public object GetService(Type serviceType) => this[serviceType];
+}
