--- test/Qowaiv.Validation.DataAnnotations.UnitTests/Models/ServiceDictionary.cs
@@ -1,6 +1,0 @@
-﻿namespace Qowaiv.Validation.DataAnnotations.UnitTests.Models;
-
-internal sealed class ServiceDictionary : Dictionary<Type, object>, IServiceProvider
-{
-    public object GetService(Type serviceType) => this[serviceType];
-}
